# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * RouteWorks API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorResponse {
  error: string;
}

export interface SplitRecipient {
  address: string;
  percentage: number;
}

export type VaultType = typeof VaultType[keyof typeof VaultType];


export const VaultType = {
  lock: 'lock',
  split: 'split',
} as const;

export type VaultStatus = typeof VaultStatus[keyof typeof VaultStatus];


export const VaultStatus = {
  active: 'active',
  paused: 'paused',
  completed: 'completed',
} as const;

export interface Vault {
  id: number;
  name: string;
  type: VaultType;
  ownerAddress: string;
  balance: number;
  status: VaultStatus;
  /** @nullable */
  description?: string | null;
  createdAt: string;
  updatedAt: string;
}

export type VaultDetailType = typeof VaultDetailType[keyof typeof VaultDetailType];


export const VaultDetailType = {
  lock: 'lock',
  split: 'split',
} as const;

export type VaultDetailStatus = typeof VaultDetailStatus[keyof typeof VaultDetailStatus];


export const VaultDetailStatus = {
  active: 'active',
  paused: 'paused',
  completed: 'completed',
} as const;

export type LockDetailsReleaseSchedule = typeof LockDetailsReleaseSchedule[keyof typeof LockDetailsReleaseSchedule];


export const LockDetailsReleaseSchedule = {
  one_time: 'one_time',
  monthly: 'monthly',
  quarterly: 'quarterly',
  annually: 'annually',
} as const;

export interface LockDetails {
  recipientAddress: string;
  amountStx: number;
  releaseSchedule: LockDetailsReleaseSchedule;
  durationMonths: number;
}

export interface VaultDetail {
  id: number;
  name: string;
  type: VaultDetailType;
  ownerAddress: string;
  balance: number;
  status: VaultDetailStatus;
  /** @nullable */
  description?: string | null;
  createdAt: string;
  updatedAt: string;
  lockDetails?: LockDetails;
  splitRecipients?: SplitRecipient[];
}

export interface VaultStats {
  totalBalance: number;
  activeVaults: number;
  upcomingPayments: number;
  totalDistributed: number;
}

export type VaultInputType = typeof VaultInputType[keyof typeof VaultInputType];


export const VaultInputType = {
  lock: 'lock',
  split: 'split',
} as const;

export interface VaultInput {
  /** @minLength 1 */
  name: string;
  type: VaultInputType;
  ownerAddress: string;
  description?: string;
  lockDetails?: LockDetails;
  splitRecipients?: SplitRecipient[];
}

export type VaultUpdateStatus = typeof VaultUpdateStatus[keyof typeof VaultUpdateStatus];


export const VaultUpdateStatus = {
  active: 'active',
  paused: 'paused',
  completed: 'completed',
} as const;

export interface VaultUpdate {
  name?: string;
  status?: VaultUpdateStatus;
  description?: string;
}

export type ActivityEventEventType = typeof ActivityEventEventType[keyof typeof ActivityEventEventType];


export const ActivityEventEventType = {
  vault_created: 'vault_created',
  funds_locked: 'funds_locked',
  payment_released: 'payment_released',
  split_executed: 'split_executed',
  recipient_added: 'recipient_added',
  vault_paused: 'vault_paused',
  vault_completed: 'vault_completed',
} as const;

export interface ActivityEvent {
  id: number;
  /** @nullable */
  vaultId?: number | null;
  /** @nullable */
  vaultName?: string | null;
  ownerAddress: string;
  eventType: ActivityEventEventType;
  description: string;
  /** @nullable */
  amount?: number | null;
  createdAt: string;
}

export type ListVaultsParams = {
/**
 * Filter by owner wallet address
 */
ownerAddress?: string;
};

export type GetVaultStatsParams = {
/**
 * Filter by owner wallet address
 */
ownerAddress?: string;
};

export type ListActivityParams = {
/**
 * Filter by owner wallet address
 */
ownerAddress?: string;
limit?: number;
};