# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: RouteWorks API specification
servers:
  - url: /api
    description: Base API path
tags:
  - name: health
    description: Health operations
  - name: vaults
    description: Flow Vault operations
  - name: activity
    description: Activity feed operations
paths:
  /healthz:
    get:
      operationId: healthCheck
      tags: [health]
      summary: Health check
      description: Returns server health status
      responses:
        "200":
          description: Healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"

  /vaults:
    get:
      operationId: listVaults
      tags: [vaults]
      summary: List all vaults for a wallet
      parameters:
        - in: query
          name: ownerAddress
          schema:
            type: string
          required: false
          description: Filter by owner wallet address
      responses:
        "200":
          description: List of vaults
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Vault"
    post:
      operationId: createVault
      tags: [vaults]
      summary: Create a new vault
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/VaultInput"
      responses:
        "201":
          description: Vault created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Vault"
        "400":
          description: Bad request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /vaults/stats:
    get:
      operationId: getVaultStats
      tags: [vaults]
      summary: Get aggregated vault statistics for a wallet
      parameters:
        - in: query
          name: ownerAddress
          schema:
            type: string
          required: false
          description: Filter by owner wallet address
      responses:
        "200":
          description: Vault statistics
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/VaultStats"

  /vaults/{id}:
    get:
      operationId: getVault
      tags: [vaults]
      summary: Get a specific vault by ID
      parameters:
        - in: path
          name: id
          schema:
            type: integer
          required: true
      responses:
        "200":
          description: Vault details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/VaultDetail"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    patch:
      operationId: updateVault
      tags: [vaults]
      summary: Update vault status or metadata
      parameters:
        - in: path
          name: id
          schema:
            type: integer
          required: true
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/VaultUpdate"
      responses:
        "200":
          description: Updated vault
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Vault"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    delete:
      operationId: deleteVault
      tags: [vaults]
      summary: Delete a vault
      parameters:
        - in: path
          name: id
          schema:
            type: integer
          required: true
      responses:
        "204":
          description: Deleted successfully
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /vaults/{id}/activity:
    get:
      operationId: getVaultActivity
      tags: [vaults]
      summary: Get activity events for a specific vault
      parameters:
        - in: path
          name: id
          schema:
            type: integer
          required: true
      responses:
        "200":
          description: List of activity events for this vault
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ActivityEvent"

  /activity:
    get:
      operationId: listActivity
      tags: [activity]
      summary: List recent activity events across all vaults
      parameters:
        - in: query
          name: ownerAddress
          schema:
            type: string
          required: false
          description: Filter by owner wallet address
        - in: query
          name: limit
          schema:
            type: integer
            default: 20
          required: false
      responses:
        "200":
          description: Recent activity events
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ActivityEvent"

components:
  schemas:
    HealthStatus:
      type: object
      properties:
        status:
          type: string
      required:
        - status

    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required:
        - error

    SplitRecipient:
      type: object
      properties:
        address:
          type: string
        percentage:
          type: number
      required:
        - address
        - percentage

    Vault:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        type:
          type: string
          enum: [lock, split]
        ownerAddress:
          type: string
        balance:
          type: number
        status:
          type: string
          enum: [active, paused, completed]
        description:
          type: ["string", "null"]
        createdAt:
          type: string
        updatedAt:
          type: string
      required:
        - id
        - name
        - type
        - ownerAddress
        - balance
        - status
        - createdAt
        - updatedAt

    VaultDetail:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        type:
          type: string
          enum: [lock, split]
        ownerAddress:
          type: string
        balance:
          type: number
        status:
          type: string
          enum: [active, paused, completed]
        description:
          type: ["string", "null"]
        createdAt:
          type: string
        updatedAt:
          type: string
        lockDetails:
          $ref: "#/components/schemas/LockDetails"
        splitRecipients:
          type: array
          items:
            $ref: "#/components/schemas/SplitRecipient"
      required:
        - id
        - name
        - type
        - ownerAddress
        - balance
        - status
        - createdAt
        - updatedAt

    LockDetails:
      type: object
      properties:
        recipientAddress:
          type: string
        amountStx:
          type: number
        releaseSchedule:
          type: string
          enum: [one_time, monthly, quarterly, annually]
        durationMonths:
          type: integer
      required:
        - recipientAddress
        - amountStx
        - releaseSchedule
        - durationMonths

    VaultStats:
      type: object
      properties:
        totalBalance:
          type: number
        activeVaults:
          type: integer
        upcomingPayments:
          type: integer
        totalDistributed:
          type: number
      required:
        - totalBalance
        - activeVaults
        - upcomingPayments
        - totalDistributed

    VaultInput:
      type: object
      properties:
        name:
          type: string
          minLength: 1
        type:
          type: string
          enum: [lock, split]
        ownerAddress:
          type: string
        description:
          type: string
        lockDetails:
          $ref: "#/components/schemas/LockDetails"
        splitRecipients:
          type: array
          items:
            $ref: "#/components/schemas/SplitRecipient"
      required:
        - name
        - type
        - ownerAddress

    VaultUpdate:
      type: object
      properties:
        name:
          type: string
        status:
          type: string
          enum: [active, paused, completed]
        description:
          type: string

    ActivityEvent:
      type: object
      properties:
        id:
          type: integer
        vaultId:
          type: ["integer", "null"]
        vaultName:
          type: ["string", "null"]
        ownerAddress:
          type: string
        eventType:
          type: string
          enum: [vault_created, funds_locked, payment_released, split_executed, recipient_added, vault_paused, vault_completed]
        description:
          type: string
        amount:
          type: ["number", "null"]
        createdAt:
          type: string
      required:
        - id
        - ownerAddress
        - eventType
        - description
        - createdAt