# lib/api-zod/src/generated/types/activityEvent.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * RouteWorks API specification
 * OpenAPI spec version: 0.1.0
 */
import type { ActivityEventEventType } from './activityEventEventType';

export interface ActivityEvent {
  id: number;
  /** @nullable */
  vaultId?: number | null;
  /** @nullable */
  vaultName?: string | null;
  ownerAddress: string;
  eventType: ActivityEventEventType;
  description: string;
  /** @nullable */
  amount?: number | null;
  createdAt: string;
}