# lib/api-zod/src/generated/types/lockDetails.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * RouteWorks API specification
 * OpenAPI spec version: 0.1.0
 */
import type { LockDetailsReleaseSchedule } from './lockDetailsReleaseSchedule';

export interface LockDetails {
  recipientAddress: string;
  amountStx: number;
  releaseSchedule: LockDetailsReleaseSchedule;
  durationMonths: number;
}