# lib/api-zod/src/generated/types/vault.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * RouteWorks API specification
 * OpenAPI spec version: 0.1.0
 */
import type { VaultStatus } from './vaultStatus';
import type { VaultType } from './vaultType';

export interface Vault {
  id: number;
  name: string;
  type: VaultType;
  ownerAddress: string;
  balance: number;
  status: VaultStatus;
  /** @nullable */
  description?: string | null;
  createdAt: string;
  updatedAt: string;
}