# lib/api-zod/src/generated/types/vaultDetail.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * RouteWorks API specification
 * OpenAPI spec version: 0.1.0
 */
import type { LockDetails } from './lockDetails';
import type { SplitRecipient } from './splitRecipient';
import type { VaultDetailStatus } from './vaultDetailStatus';
import type { VaultDetailType } from './vaultDetailType';

export interface VaultDetail {
  id: number;
  name: string;
  type: VaultDetailType;
  ownerAddress: string;
  balance: number;
  status: VaultDetailStatus;
  /** @nullable */
  description?: string | null;
  createdAt: string;
  updatedAt: string;
  lockDetails?: LockDetails;
  splitRecipients?: SplitRecipient[];
}