# lib/api-zod/src/generated/types/vaultInput.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * RouteWorks API specification
 * OpenAPI spec version: 0.1.0
 */
import type { LockDetails } from './lockDetails';
import type { SplitRecipient } from './splitRecipient';
import type { VaultInputType } from './vaultInputType';

export interface VaultInput {
  /** @minLength 1 */
  name: string;
  type: VaultInputType;
  ownerAddress: string;
  description?: string;
  lockDetails?: LockDetails;
  splitRecipients?: SplitRecipient[];
}