# lib/api-zod/src/generated/types/vaultStats.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * RouteWorks API specification
 * OpenAPI spec version: 0.1.0
 */

export interface VaultStats {
  totalBalance: number;
  activeVaults: number;
  upcomingPayments: number;
  totalDistributed: number;
}